refactor(#80): 채널 채팅 서버 데이터 페칭 _data 폴더로 이동
