+++ src/app/channel/chat/page.tsx
@@ -1,7 +1,7 @@
 // 채널 채팅 설정 페이지를 렌더링합니다.
 import { ChannelChatPageContent } from "@/components/channel/chat/channel-chat-page-content";
 
-import { getChannelChatSnapshot } from "./data";
+import { getChannelChatSnapshot } from "./_data/channel-chat-data";
 
 export default async function ChannelChatPage() {
   const result = await getChannelChatSnapshot();
